pyproject.toml
[tool.poetry]
name = "kethaa"
version = "0.0.1"
description = "An Ethereum Account Abstraction on Starknet used to interact with the Kakarot ZKEVM"
authors = ["Flydexo"]
license = "MIT"
readme = "README.md"
packages = [{include = "kakarot"}]

[tool.poetry.dependencies]
python = "3.9.1"
cairo-lang = "0.10.3"
openzeppelin-cairo-contracts = "^0.6.0"
amarna = "^0.1.5"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.11.4"
pytest-cov = "^4.0.0"
starknet-py = "^0.12.0a0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
